requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
cssselect==1.0.1
defusedxml==0.5.0
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
django-libsass==0.7
django-modelcluster==3.0.1
django-taggit==0.22.0
django-treebeard==4.1.0
djangorestframework==3.6.2
et-xmlfile==1.0.1
grab==0.6.35
html5lib==0.999999999
hurry.filesize==0.9
jdcal==1.3
libsass==0.12.3
lxml==3.7.2
olefile==0.44
openpyxl==2.4.2
packaging==16.8
Pillow==4.0.0
psycopg2==2.6
pycurl==7.43.0
pyparsing==2.1.10
python3-wget==0.0.2b1
pytils==0.3
pytz==2016.10
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
selection==0.0.13
six==1.10.0
steamodd==4.21
Unidecode==0.4.20
user-agent==0.1.5
wagtail==1.9
webencodings==0.5
weblib==0.1.24
wget==3.2
Willow==0.4
dj-database-url==0.4.1
whitenoise==2.0.6
gunicorn==19.6.0